Move files around for implementing persistence
