+++ crates/apps/looper/looper-processor/src/audio/mod.rs
@@ -1,0 +1,6 @@
+pub mod loop_quantization;
+pub mod midi_map;
+pub mod multi_track_looper;
+pub mod processor;
+pub mod sequencer;
+pub mod time_info_provider;
