+++ crates/apps/looper/looper-processor/src/audio/multi_track_looper/lfo_processor.rs
