+++ crates/apps/looper/looper-processor/src/audio/multi_track_looper/metrics/audio_processor_metrics/mod.rs
@@ -1,8 +1,8 @@
-use basedrop::Shared;
 use std::cmp::Ordering;
 use std::collections::VecDeque;
-
 use std::time::Duration;
+
+use basedrop::Shared;
 
 pub use audio_thread::*;
