+++ crates/apps/looper/looper-processor/src/audio/multi_track_looper/slice_worker/mod.rs
@@ -14,7 +14,7 @@ use audio_processor_traits::{
     AudioBuffer, AudioProcessorSettings, OwnedAudioBuffer, VecAudioBuffer,
 };
 
-use crate::processor::handle::LooperClipRef;
+use crate::audio::processor::handle::LooperClipRef;
 
 #[derive(Clone)]
 pub struct SliceResult {
