+++ crates/apps/looper/looper-processor/src/audio/multi_track_looper/tempo_estimation/mod.rs
