+++ crates/apps/looper/looper-processor/src/audio/processor/handle/scratch_pad.rs
