+++ crates/apps/plugin-host/plugin-host-lib/src/lib.rs
@@ -1,6 +1,6 @@
+pub use actix_system_threads as actor_system;
 pub use audio_io::test_plugin_host::TestPluginHost;
 
-pub mod actor_system;
 pub mod audio_io;
 pub mod commands;
 pub mod processors;
