+++ crates/augmented/application/actix-system-threads/README.md
@@ -1,0 +1,3 @@
+# actix-system-threads
+
+This is a wrapper struct that provides a single global dispatcher into several actix `Arbiter` threads.
